requirements.txt
flask_bootstrap==3.3.7.1
apscheduler==3.10.4
pytz
Bootstrap-Flask==2.2.0
flask_sqlalchemy==3.0.3
flask_migrate==4.0.0
flask_login==0.6.2
flask[async]==2.3.2
python-dotenv==1.0.0
requests==2.25.1
google-api-python-client==2.87.0
oauth2client==4.1.3
google-auth-httplib2==0.1.0
requests-oauthlib==1.3.1
oauthlib==3.2.2
google-auth-oauthlib==1.0.0
unidecode==1.3.6